Windows terminal support
