+++ go/client/color.go
@@ -89,16 +89,6 @@ func colorByteSequence(code int) []byte {
 func (cp CodePair) OpenBytes() []byte  { return colorByteSequence(cp.Open) }
 func (cp CodePair) CloseBytes() []byte { return colorByteSequence(cp.Close) }
 
-func HasColor() bool {
-	// TODO Color should be based on whether log format supports it
-	logFormatHasColor := map[string]bool{
-		"":        true,
-		"default": true,
-		"fancy":   true,
-	}
-	return logFormatHasColor[G.Env.GetLogFormat()]
-}
-
 func ColorOpen(which string) []byte {
 	if !HasColor() {
 		return nil
