+++ go/minterm/minterm_windows.go
@@ -1,0 +1,24 @@
+// +build windows
+
+package minterm
+
+import (
+	"golang.org/x/crypto/ssh/terminal"
+	"os"
+)
+
+func (m *MinTerm) open() error {
+	fout, err := os.OpenFile("CONOUT$", os.O_RDWR, 0)
+	if err != nil {
+		return err
+	}
+
+	m.out = fout
+	fd := int(m.out.Fd())
+	w, h, err := terminal.GetSize(fd)
+	if err != nil {
+		return err
+	}
+	m.width, m.height = w, h
+	return nil
+}
